move remaning interfaces to more specific files
