+++ src/components/api/const/interfaces/registrationInterfaces.tsx
@@ -1,0 +1,5 @@
+export interface RegisterFormInputs {
+  name: string;
+  email: string;
+  password: string;
+}
